mageia-accounts.gsb: add transactions from 2012
Added using data from 2012 pdf report downloaded from :
http://www.mageia.org/en/about/reports/2012/mageia_rapport_financier_2012_fr.pdf

There are a some errors in this report.

Total amount for donations received is wrong, on page 7 :

    615.00 (check)
 + 3440.80 (transfer)
 + 5946.28 (paypal)
 +  160.00 (cash)
 ----------
  10162.08

But the total in the report says 10072.08. It should be 10162.08.
It's also wrong on page 5, and on the report on the web site.

Total balance on page 4 and on the web site is wrong. It says 8423.07
euros on 2012/12/31 but :

    8166.37 (balance from 2011/12/31)
 + 10162.08 (income for 2012)
 - 10766.69 (expenses for 2012)
 -----------
    7561.76

So correct balance should be 7561.76 (which is also what we have in
grisbi) instead of 8423.07. Unless there is some income that is missing
in the report. With wrong donations total (10072.08 instead of 10162.08)
the balance would be 7471.76, so this is 2 different errors.

git-svn-id: svn+ssh://svn.mageia.org/svn/treasurer@4 54c5794f-510f-4faa-a1d1-3574049f8189